Add new threading-based udpsmoke implementation
Add new threading-based udpsmoke implementation.
The new protocol is optimized and is not 100% compatible with the
old udpsmoke-async.
The threading udpsmoke has NO EXTERNAL libs dependency. Just
Python 3.5+ is needed and it runs on new Linux and FreeBSDs for
both IPv4 and IPv6.
There is hardcoded very basic Prometheus exporter (not using proper
library).