Add support for resolving hostnames on start
Add support for resolving DNS hostnames in the tartget list on start.
The hostnames are first tested if the hostname is actually an IPv4 or
IP6 address. If not we attmpt a resolution in DNS as IPv6 and if
unsuccessful then as IPv4. If nothing works the startup procedure fails.